added security

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use App\Post;
 
 class PostsController extends Controller
 {
+        /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -83,6 +93,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //CHeck for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts')->with('error','Unathoritzdoage
+            ');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -119,6 +135,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts')->with('error','Unauthorized Page');
+        }
+
         $post->delete();
 
         return redirect('posts')->with('success','Post Removed');
